KZ80 initial board
This is the initial board that uses a 555 with a pot to control the click frequency.  To reset, the reset wire needs to be connected to ground for two click cycles and you will the 8-bit address display clear.  Then, reconnect reset wire to power to start cpu at #0000.  Note, thr cmos Z84C cpu I'm using runs at a slow clock speed as the pot is set to cycle every 2 seconds.

Let me know what you think,

Diana